ml/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
spacy>=3.8.0
PyPDF2>=3.0.0
python-docx>=1.1.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.26.0
python-multipart>=0.0.6
